+++ asapdiscovery-modeling/asapdiscovery/modeling/tests/test_schema.py
@@ -4,7 +4,6 @@ import pydantic
 import pytest
 from asapdiscovery.data.openeye import oechem
 from asapdiscovery.data.testing.test_resources import fetch_test_file
-from asapdiscovery.modeling.modeling import save_design_unit
 from asapdiscovery.modeling.schema import MoleculeFilter
